add restart app

+++ src/main/java/Factorial.java
@@ -7,7 +7,7 @@ public class Factorial {
     public static long factorial(int number) {
         long result;
         if (number > 1)
-            return number * factorial(number - 1);
+            return (long)number * factorial(number - 1);
         else if (number==1 || number ==0)
             return 1;
         else
